- Made objects that fall under gravity to also take into account their mass. - Fixed most compiling warnings under MSVC. - Made flying monsters take into account monsters above or below them when changing their height. - Made objects that are on other objects to respect the floor height, in case the sector changes it's height. - Added some performance CVars to avoid drawing shadows or lights on models.

+++ utils/master/main.cpp
@@ -269,7 +269,7 @@ int main(int argc, const char** argv)
 
 		//	Clean up list from old records.
 		time_t CurTime = time(0);
-		for (size_t i = 0; i < SrvList.Num(); i++)
+		for (size_t i = 0; i < (unsigned int)SrvList.Num(); i++)
 		{
 			if (CurTime - SrvList[i].Time > 15 * 60)
 			{
